Update README to discourage use of wildcards in BP_JVM_VERSION
Wildcard values like 11.* will stil work as before, but they must match the version of OpenJDK that a realeas of GraalVM is based on rather than matching the GraalVM version. This is potentially confusing. Encouraging users to use JVM spec version like 8 and 11 is more intuitive and less prone to break when on buildpack upgrades.

Signed-off-by: Emily Casey <ecasey@vmware.com>