This "uses" the helpers module so that it is retained by `go mod`
This is not needed by the application but our build script will build the helper cmd and that does require this library. Because only the build script requires it, the `go mod` system does not know about this requirement and will prune the required dependencies from `go.mod` and `go.sum`. This causes the build to break if you run `go mod tidy`. Unforunately, when Dependabot submits a PR to bump dependencies, it runs `go mod tidy`, and so every Dependabot PR breaks our project which requires a manual PR to fix.

This change should resolve that issue because it will force `go mod` to retain the required modules for our build script.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>